requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
openpyxl
python-jose[cryptography]
bcrypt